good boost

+++ cpp-tutorial/myexamples/boost/main.cc
@@ -19,7 +19,7 @@ int main() {
     vector<string> result;
     boost::split(result, input, boost::is_any_of("\t"));
 
-    for (int i = 0; i < result.size(); i++) {
+    for (size_t i = 0; i < result.size(); i++) {
         std::cout << result[i] << std::endl;
     }
     return 0;
